backend/requirements.txt
uvicorn
fastapi
python-multipart
SQLAlchemy
psycopg2-binary
openai
python-dotenv